requirements.txt
streamlit
openai
llama-index
llama-index-llms-openai
nltk
pypdf
llama-index-readers-microsoft-sharepoint
python-dotenv